chg: restructure import files
